requirements.txt
torch>=2.0.0
transformers>=4.30.0
clip-by-openai
Pillow>=9.0.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.28.0
